Modified integration test
