+++ src/test/java/com/blackducksoftware/integration/hub/alert/channel/slack/SlackChannelTestIT.java
@@ -14,16 +14,18 @@ package com.blackducksoftware.integration.hub.alert.channel.slack;
 import org.junit.Assert;
 import org.junit.Test;
 
+import com.blackducksoftware.integration.hub.alert.channel.RestChannelTest;
 import com.blackducksoftware.integration.hub.alert.datasource.entity.SlackConfigEntity;
-import com.google.gson.Gson;
 
-public class SlackChannelTestIT {
+public class SlackChannelTestIT extends RestChannelTest {
 
     @Test
     public void sendMessageTestIT() {
-        final Gson gson = new Gson();
         final SlackChannel slackChannel = new SlackChannel(gson, null);
-        final SlackConfigEntity config = new SlackConfigEntity("#alert", "webhookbot", "https://hooks.slack.com/services/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");
+        final String roomName = properties.getProperty("slack.room.name");
+        final String username = properties.getProperty("slack.username");
+        final String webHook = properties.getProperty("slack.web.hook");
+        final SlackConfigEntity config = new SlackConfigEntity(roomName, username, webHook);
         final String actual = slackChannel.testMessage(config);
         final String expected = "200";
         Assert.assertEquals(expected, actual);
